requirements.txt
Appium-Python-Client==2.1.4
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chromedriver==2.24.1
click==7.1.2
colorama==0.4.4
cors==1.0.1
cryptography==36.0.1
decorator==5.1.1
Deprecated==1.2.13
filelock==3.6.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
get-geckodriver==0.2.1
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-resources==5.6.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
MarkupSafe==1.1.1
modcall==0.1.0
module==0.0.4
mysql==0.0.2
mysqlclient==2.0.3
namespace==0.1.4
outcome==1.1.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.0.1
pytest-html==3.1.1
pytest-metadata==2.0.1
Python-EasyConfig==0.1.7
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requests-file==1.5.1
Resource==0.2.1
selenium==3.3.0
self==2020.12.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
tldextract==3.2.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==1.0.1
wrapt==1.14.0
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
